Moved to include/ai
